Create new status for test timeouts
